extension
Steam User Manager is an extension that works as an auxiliary tool for Steam. Its functionality includes the ability to quickly…
Steam User Manager is an extension that works as an auxiliary tool for Steam. Its functionality includes the ability to quickly switch between accounts, work with multiple accounts at the same time. The ability to receive a 2fa code and confirm offers directly from the browser.